[refactoring] step2: value class 사용
