+++ src/main/kotlin/blackjack/ui/printer/ParticipantCardsPrinter.kt
@@ -1,0 +1,10 @@
+package blackjack.ui.printer
+
+import blackjack.entity.ParticipantCards
+
+@JvmInline
+value class ParticipantCardsPrinter(private val participantCards: ParticipantCards) {
+    fun print(): String {
+        return participantCards.joinToString { card -> CardPrinter(card).print() }
+    }
+}
